+++ udata_front/theme/gouvfr/assets/js/types.ts
@@ -180,9 +180,16 @@ export type RefusedMembershipRequest = MembershipRequest & {
 
 export type MemberRole = "admin" | "editor";
 
+// In org endpoint we get these two private information if we have edit rights on the org.
+export type MemberUser = User & {
+  email: string | null;
+  last_login_at: string | null;
+}
+
 export type Member = {
   role: MemberRole;
-  user: User;
+  user: MemberUser;
+  since: string;
 };
 
 export type EditingMember = Member & {